fix: extract API resolver code in separate file
