+++ packages/api-page-builder/src/graphql/graphql/pages.gql.ts
@@ -8,6 +8,7 @@ import { GraphQLSchemaPlugin } from "@webiny/handler-graphql/types";
 import { Page, PbContext, PageSecurityPermission } from "~/types";
 import WebinyError from "@webiny/error";
 import resolve from "./utils/resolve";
+import resolvePageBlocks from "./utils/resolvePageBlocks";
 import { createPageSettingsGraphQL } from "./pages/pageSettings";
 import { fetchEmbed, findProvider } from "./pages/oEmbed";
 import lodashGet from "lodash/get";
@@ -283,46 +284,7 @@ const createBasePageGraphQL = (): GraphQLSchemaPlugin<PbContext> => {
                         if (!page.content?.elements) {
                             return page.content;
                         }
-
-                        const blocks = [];
-                        for (const pageBlock of page.content?.elements) {
-                            const blockId = pageBlock.data?.blockId;
-                            // If block has blockId, then it is reference block and we need to get elements for it
-                            if (blockId) {
-                                const blockData = await context.pageBuilder.getPageBlock(blockId);
-                                // We check if block has variable values set on the page and use them in priority over ones,
-                                // that are set in blockEditor
-                                const blockDataVariables =
-                                    blockData?.content?.data?.variables || [];
-                                const variables = blockDataVariables.map(
-                                    (blockDataVariable: any) => {
-                                        const value =
-                                            pageBlock.data?.variables?.find(
-                                                (variable: any) =>
-                                                    variable.varRef === blockDataVariable.varRef
-                                            )?.value || blockDataVariable.value;
-
-                                        return {
-                                            ...blockDataVariable,
-                                            value
-                                        };
-                                    }
-                                );
-
-                                blocks.push({
-                                    ...pageBlock,
-                                    data: {
-                                        blockId,
-                                        ...blockData?.content?.data,
-                                        variables
-                                    },
-                                    elements: blockData?.content?.elements || []
-                                });
-                            } else {
-                                blocks.push(pageBlock);
-                            }
-                        }
-
+                        const blocks = await resolvePageBlocks(page, context);
                         return { ...page.content, elements: blocks };
                     }
                 },
